requirements.txt
argparse==1.2.1
psycopg2==2.6.1
django-filter==0.7
djangorestframework==2.3.14
django-rest-swagger==0.2.0
wsgiref==0.1.2
Django==1.7
